Add files via upload
This Document presents the steps for generating and utilising the models of “classification of types of 
transactions” 
The data flow architecture for the solution is given below: 




The primary goal here is to generate the labels first and then train a supervised model using these 
generated data to classify the types of transaction. 
We have been given a non-labelled data of bank transactions, these data have various levels of 
transactions in it and it may happen that we will get completely new input data in future. The goal is 
to minimize or (completely get rid of expensive models like OpenAI to get the final prediction) and 
use comprehensive supervised classification model to predict the old as well as completely new kind 
of input data into its respective class 
1. Process of Labelling the dataset:   
we have almost 5000, samples (raw) after cleaning the provided input dataset.   These data samples 
are not labelled and its quite hard to tag them with label manually, here we have introduced the first 
block of the solution: use of OpenAI to label the raw dataset  
As you can see that using prompts, we have passed the raw input (cleaned) data into the OpenAI 
LLM and asked them to given them the label as per its description. Below is the snapshot of the 
labels provided by openai for some of the complex description: 
2.Validate the labels by OpenAI:  
In this simple step, we manually picked random samples from the OpenAI and validated the labels, 
the temperature and Top-p hyperparameters were set to minimum and the prompts were optimally 
designed therefore we found that almost each description was labelled correctly  
Below are the total labels we used: 
1.Personal Expense: Education, Entertainment, Travel etc 
2.Online Purchase: Online shopping, Subscriptions, mandates etc 
3.Food Bill: Restaurant bill, Online Food Bill etc 
4. Card Payment: transaction where the details of “what” transaction are not defined but card details 
are mentioned 
5. Bank to Bank /Deposits - Personal transaction: Transaction done to other banks or person 
Once we have validated the transaction type on selected data samples, we are now ready to develop 
machine learning model  
3.Model Generation: 
Although I have designed 2 different ways to predict the type of transaction, the best way comes out 
as the 2nd way of “Use of Embeddings with PCA”, so in this document I would be explaining the 
working of this model. 
Flowchart of “Embedding model with PCA.” 
The input data is first converted into embedding vector (Numerical representation of the description) 
, This embedding or vector is of dimension 768*1 , to compress this size and reduce the dimension 
we used PCA , the usage of PCA was subjected to one constrain : “we should retain at least 80% of 
the 768 dimension’s information” 
The optimal number of Principal components which were retaining almost ~82% of original 
information was developed and we successfully converted the information from 768*1 Dimension to 
50*1 dimension 
We then used this 50*1 data to develop a ML model, below are the summary of ML model 
developed on this dataset  
Note : All the models are developed using Gridsearch with analysis of different parameters like 
mean_test_score , mean_train_score to validate if the models are overfitting or underfitting , thus 
these developed models are reliable  
1.SVC : Support vector Machine Classifier works very good on non-linear classification problem , 
given that the nature of data and high dimensions , this algorithm was natural first choice , the 
algorithm was build on following hyperparameters : 
The results was: Accuracy of  ~85% 
2.Decision Tree: This rule-based classifier is again a good choice for the non-linear Multiclass 
classification problem : 
This algorithm was build on following hyperparameters : 
The results were : Accuracy of ~ 71.2% 
3. Randomforest classifier : This is a type of bagging model , although the size / number of samples 
of data are quite small , we tried this model with following hyperparameters : 
The results were : Accuracy of ~82% 
4.Nueral Network (Multi layered Perceptron) : This Advancend Deep learning model was applied on 
following hyperparameters : 
(For better picture please refer to training notebook) 
The results were: Accuracy of ~79.5%